Move secrets main() to direktiv main().
