+++ cmd/secrets/entry.go
@@ -1,4 +1,4 @@
-package main
+package secrets
 
 import (
 	"log"
@@ -9,7 +9,7 @@ import (
 	"github.com/direktiv/direktiv/pkg/secrets"
 )
 
-func main() {
+func RunApplication() {
 
 	backend := "db"
